Move alteration to operations/cast, fix tests.
