+++ db/tests/types/operations/test_cast.py
@@ -12,7 +12,7 @@ from db.columns.operations.select import get_column_default
 from db.columns.operations.alter import alter_column_type
 from db.tables.operations.select import get_oid_from_table
 from db.tests.types import fixtures
-from db.types import alteration
+from db.types.operations import cast as cast_operations
 from db.types.base import PostgresType, MathesarCustomType, get_qualified_name, get_available_types
 
 
@@ -360,7 +360,7 @@ MASTER_DB_TYPE_MAP_SPEC = {
 
 
 def test_get_alter_column_types_with_custom_engine(engine_with_types):
-    type_dict = alteration.get_supported_alter_column_types(engine_with_types)
+    type_dict = cast_operations.get_supported_alter_column_types(engine_with_types)
     assert all(
         [
             type_ in type_dict.values()
@@ -370,7 +370,7 @@ def test_get_alter_column_types_with_custom_engine(engine_with_types):
 
 
 def test_get_alter_column_types_with_unfriendly_names(engine_with_types):
-    type_dict = alteration.get_supported_alter_column_types(
+    type_dict = cast_operations.get_supported_alter_column_types(
         engine_with_types, friendly_names=False
     )
     assert all(
@@ -639,7 +639,7 @@ def test_get_column_cast_expression_unchanged(engine_with_types):
     target_type = "numeric"
     col_name = "my_column"
     column = Column(col_name, Numeric)
-    cast_expr = alteration.get_column_cast_expression(
+    cast_expr = cast_operations.get_column_cast_expression(
         column, target_type, engine_with_types
     )
     assert cast_expr == column
@@ -649,7 +649,7 @@ def test_get_column_cast_expression_change(engine_with_types):
     target_type = "boolean"
     col_name = "my_column"
     column = Column(col_name, Numeric)
-    cast_expr = alteration.get_column_cast_expression(
+    cast_expr = cast_operations.get_column_cast_expression(
         column, target_type, engine_with_types
     )
     assert str(cast_expr) == f"mathesar_types.cast_to_boolean({col_name})"
@@ -659,7 +659,7 @@ def test_get_column_cast_expression_change_quotes(engine_with_types):
     target_type = "boolean"
     col_name = "A Column Needing Quotes"
     column = Column(col_name, Numeric)
-    cast_expr = alteration.get_column_cast_expression(
+    cast_expr = cast_operations.get_column_cast_expression(
         column, target_type, engine_with_types
     )
     assert str(cast_expr) == f'mathesar_types.cast_to_boolean("{col_name}")'
@@ -668,8 +668,8 @@ def test_get_column_cast_expression_change_quotes(engine_with_types):
 def test_get_column_cast_expression_unsupported(engine_with_types):
     target_type = "this_type_does_not_exist"
     column = Column("colname", Numeric)
-    with pytest.raises(alteration.UnsupportedTypeException):
-        alteration.get_column_cast_expression(
+    with pytest.raises(cast_operations.UnsupportedTypeException):
+        cast_operations.get_column_cast_expression(
             column, target_type, engine_with_types
         )
 
@@ -692,7 +692,7 @@ def test_get_column_cast_expression_numeric_options(
 ):
     target_type = "numeric"
     column = Column("colname", type_)
-    cast_expr = alteration.get_column_cast_expression(
+    cast_expr = cast_operations.get_column_cast_expression(
         column, target_type, engine_with_types, type_options=options,
     )
     assert str(cast_expr) == expect_cast_expr
@@ -706,7 +706,7 @@ expect_cast_tuples = [
 
 @pytest.mark.parametrize("source_type,expect_target_types", expect_cast_tuples)
 def test_get_full_cast_map(engine_with_types, source_type, expect_target_types):
-    actual_cast_map = alteration.get_full_cast_map(engine_with_types)
+    actual_cast_map = cast_operations.get_full_cast_map(engine_with_types)
     actual_target_types = actual_cast_map[source_type]
     assert len(actual_target_types) == len(expect_target_types)
     assert sorted(actual_target_types) == sorted(expect_target_types)
