+++ db/types/exceptions.py
@@ -1,0 +1,2 @@
+class UnsupportedTypeException(Exception):
+    pass
