+++ db/types/install.py
@@ -1,11 +1,12 @@
-from db.types import base, alteration, email
+from db.types import base, email
 from db.schemas.operations.create import create_schema
+from db.types.operations.cast import install_all_casts
 
 
 def install_mathesar_on_database(engine):
     create_type_schema(engine)
     email.install(engine)
-    alteration.install_all_casts(engine)
+    install_all_casts(engine)
 
 
 def create_type_schema(engine):
